flake: inputs: remove flake-utils
Why? You don't need it: https://ayats.org/blog/no-flake-utils/

As for `systems`, this is a useful abstraction that allows you to
specify a list of systems either as a file (`url =
"path:./flake.systems.nix"`) or by the default repository as shown.

Read up: https://github.com/nix-systems/nix-systems

This was the biggest selling point of `flake-utils`, but `nix-systems`
is a simpler (and more easily composable) abstraction.